Finished Navigation Bar Layout.
I have finished the layout and the structure of the side navigation bar and added some what of a basic responsiveness to it. I have also created a redux store, one of which's features is to manage parts of the styling of the application. One of this parts is the toggle state of the navigation bar.