Routines Widget.
I have created the routine widget that is responsible for the management of all the created routines by the user. I have implemented some dummy data so that I can work on the filtering as well. The tag filtering is now finished and works perfectly fine. The widget is static though, and still have not any actual responsiveness.