Working on the Sign Up Form.
I have started the development of the sign up form. I have made a lot of progress - created the layout of the form at the front end with all the fetching logic between the client and the server and also created some of the RESTAPI at the back end responsible for the sign up requests. There are lot more modification that are coming next after this version.