Finished Log In and Sign In Forms.
I have now finished and created working log in and sign in forms both on the front and back end. I have the layout of the forms (still not responsive) with all the validation management based on the responses created by the RestAPIs at the server side of the application.