Adding Exercise Options.
I have added the options menu for each of the currant exercises that allow you either to remove the exercise completely or replace it with a new one. I have also implemented the feature of being able to get summary about the exercise the user has already added by clicking on the image or the name.