Layout Configurations.
I have fixed some layout bugs of the workout page and constructed it better so the widget are more manageable later on when I am working on the responsiveness of the application. I have also added the workout pad which is available when a workout session is active and via it the user is able to access the tracker itself or minimize it when he want to browse through the application.