New Routine Form.
I have added the new routine form that is responsible for the creation of a new routine for the user that later on he can use as a template for his workout.

The creation process included development of an additional RestAPI on the backend as well as some modification on the front end related to fixing some encountered bugs and lay outing the form itself.