Workout Tracking.
I have implemented the functionality of starting a workout, tracking it and finally saving it as a user data at the database when the user finishes their workout. When that happens the user is greeted with a message showing his progress (basic information).
This process included the development of a RESTAPI at the back end as well as some modifications to the redux store at the front end, the communication between the two sides and adjusting some bugs.