Calendar Widget.
I have added all the functionality of the calendar widget. Now the user is able to see the dates when he has workout and access the history of his workouts by clicking on a given date. By doing so a modal is open that is giving simple preview of what has the user trained that date. If there were multiple workouts than all of them are presented.