Activity Sessions.
I have created the functionality of tracking activity session by clicking the categories buttons. By doing so, the user activates a different kind of workout that tracks the calories he is burning based on the selected activity and the total kilograms of the user for the average person. The activities can be submitted and saved to the user's history which allows them to be reviewed later in the calendar.

In the feature several uncaught bugs can be encountered with the provided changes when it comes to the feature based on the lack of testing.