Health Essentials Data Management.
I have finished the main purpose of the health essentials page. I have created a couple of APIs and modified some old ones that are responsible for the management of the user's essentials page.

I have also refactored some of the code both on the back and the front end and fixed a couple of bugs.