Explore Page.
I have build up from the ground the explore page with all of its functionality, components, layouts and additional pages. This process included quite a development both on the back end and the front end.

I have also started working on the layout of the statistics page, but it is in its early stages of development, so I can't be certain if the idea behind it will be official.