Landing Page & Responsiveness.
The goal of this patch was to develop the responsiveness of the application on as much as possible screens of different devices. I have also added a simple Landing Page (on which later on will be added a footer as well) that allows the user to navigate to the application and to get a general understanding of the product. The landing page is also responsive.

I have also fixed some crucial bugs when it comes to adding and removing sets in the workout tracker from the application. A few more issues were fixed during the process of development.